Changed file structure of components and seperate the animated gradient in different files
