+++ src/components/screens/SchoolScheduleTakePicture.js
